requirements.txt
coverage>=4.5.3
Django>=2.2.3
djangorestframework>=3.9.4
ipython>=7.6.1
psycopg2-binary>=2.8.3
